+++ src/Polars/Internal/Raw.hs
@@ -64,6 +64,7 @@ module Polars.Internal.Raw
     , phs_expr_slice
     , phs_expr_sort_by
     , phs_expr_string_function
+    , phs_expr_string_function_i64
     , phs_expr_ternary
     , phs_expr_unary
     , phs_expr_unary_i64
@@ -404,3 +405,6 @@ foreign import ccall unsafe "phs_expr_over"
 
 foreign import ccall unsafe "phs_expr_string_function"
     phs_expr_string_function :: CInt -> Ptr RawExpr -> Ptr (Ptr RawExpr) -> CSize -> Ptr (Ptr RawExpr) -> Ptr (Ptr RawError) -> IO CInt
+
+foreign import ccall unsafe "phs_expr_string_function_i64"
+    phs_expr_string_function_i64 :: CInt -> CLLong -> Ptr RawExpr -> Ptr (Ptr RawExpr) -> CSize -> Ptr (Ptr RawExpr) -> Ptr (Ptr RawError) -> IO CInt